Extract base class from Invenio deposit plugin
